Update weirdalgorithm.cpp to handle larger input values and add problem reference. problem pass

+++ cses/weirdalgorithm.cpp
@@ -1,3 +1,6 @@
+// pass
+// problem url: https://cses.fi/problemset/task/1068
+
 #include <iostream>
 
 using namespace std;
@@ -6,7 +9,7 @@ int main() {
     ios::sync_with_stdio(false);
     cin.tie(nullptr);
 
-    int n;
+    long long n;
     cin >> n;
 
     while (n != 1) {
